Refactor MutantRunner to separate TestRunner initialization
