+++ core/src/main/scala/stryker4s/run/Stryker4sRunner.scala
@@ -1,6 +1,8 @@
 package stryker4s.run
 
-import cats.effect.{ContextShift, IO}
+import java.nio.file.Path
+
+import cats.effect.{ContextShift, IO, Resource, Timer}
 import cats.syntax.all._
 import stryker4s.Stryker4s
 import stryker4s.config._
@@ -9,14 +11,14 @@ import stryker4s.log.Logger
 import stryker4s.mutants.Mutator
 import stryker4s.mutants.applymutants.ActiveMutationContext.ActiveMutationContext
 import stryker4s.mutants.applymutants.{MatchBuilder, StatementTransformer}
-import stryker4s.mutants.findmutants.{FileCollector, MutantFinder, MutantMatcher, SourceCollector}
+import stryker4s.mutants.findmutants.{FileCollector, MutantFinder, MutantMatcher}
 import stryker4s.report._
 import stryker4s.report.dashboard.DashboardConfigProvider
 import stryker4s.run.process.ProcessRunner
 import stryker4s.run.threshold.ScoreStatus
 import sttp.client.asynchttpclient.cats.AsyncHttpClientCatsBackend
 
-abstract class Stryker4sRunner(implicit log: Logger, cs: ContextShift[IO]) {
+abstract class Stryker4sRunner(implicit log: Logger, cs: ContextShift[IO], timer: Timer[IO]) {
   def run(): IO[ScoreStatus] = {
     implicit val config: Config = ConfigReader.readConfig()
 
@@ -30,7 +32,7 @@ abstract class Stryker4sRunner(implicit log: Logger, cs: ContextShift[IO]) {
           new StatementTransformer,
           resolveMatchBuilder
         ),
-        resolveRunner(collector, new AggregateReporter(reporters))
+        new MutantRunner(resolveTestRunner(_), collector, new AggregateReporter(reporters))
       )
       stryker4s.run()
     }
@@ -48,9 +50,9 @@ abstract class Stryker4sRunner(implicit log: Logger, cs: ContextShift[IO]) {
           }
     }
 
-  def resolveRunner(collector: SourceCollector, reporter: Reporter)(implicit config: Config): MutantRunner
-
   def resolveMatchBuilder(implicit config: Config): MatchBuilder = new MatchBuilder(mutationActivation)
+
+  def resolveTestRunner(tmpDir: Path)(implicit config: Config): Resource[IO, stryker4s.run.TestRunner]
 
   def mutationActivation(implicit config: Config): ActiveMutationContext
 }
